Windows build fixed

+++ benches/variantbenches.cpp
@@ -19,12 +19,12 @@ void e172_Variant_ts_bar(int value)
 std::pair<int64_t, int64_t> e172::benches::VariantBenches::speedTest(size_t count)
 {
     ElapsedTimer t;
-    for(size_t i = 0; i < count; ++i) {
+    for(int i = 0; i < count; ++i) {
         e172_Variant_ts_foo(i);
     }
     const auto t0 = t.elapsed();
     t.reset();
-    for(size_t i = 0; i < count; ++i) {
+    for(int i = 0; i < count; ++i) {
         e172_Variant_ts_bar(i);
     }
     const auto t1 = t.elapsed();
@@ -48,12 +48,11 @@ double e172::benches::VariantBenches::speedTest()
 
 void e172::benches::VariantBenches::banchmark()
 {
-    size_t sum = 0;
+    double sum = 0;
     for(size_t i = 0; i < 100; ++i) {
         const auto s = speedTest();
         Debug::print("[", i, "%] e172::VariantTest::speedTest:", s);
         sum += s;
     }
-    Debug::print("e172::VariantTest::speedTest (average):", sum / 100);
+    Debug::print("e172::VariantTest::speedTest (average):", static_cast<std::size_t>(sum / 100.));
 }
-
